condensed files
